Complete ECOM-10-Occasions with unit test coverage
