+++ lib/core/constants/app_images.dart
@@ -7,4 +7,6 @@ abstract final class AppImages {
   static const String logo = '$_rootImages/Logo_flower.png';
   static const String productTest = '$_rootImages/prodcutTest.png';
   static const String productTestImage = '$_rootImages/prodeuctTestImage.png';
+  static const String fakeImgUrl =
+      "https://flower.elevateegy.com/uploads/fefa790a-f0c1-42a0-8699-34e8fc065812-cover_image.png";
 }
