+++ lib/presentation/occasion/views/widgets/occasion_view_body.dart
@@ -1,11 +1,10 @@
 import 'package:elevate_ecommerce_app/core/constants/app_colors.dart';
 import 'package:elevate_ecommerce_app/core/di/di.dart';
-import 'package:elevate_ecommerce_app/core/router/route_names.dart';
-import 'package:elevate_ecommerce_app/core/utils/widgets/custom_product_items.dart';
 import 'package:elevate_ecommerce_app/generated/l10n.dart';
 import 'package:elevate_ecommerce_app/presentation/occasion/view_models/occasion_events.dart';
 import 'package:elevate_ecommerce_app/presentation/occasion/view_models/occasion_states.dart';
 import 'package:elevate_ecommerce_app/presentation/occasion/view_models/occasion_view_model.dart';
+import 'package:elevate_ecommerce_app/presentation/occasion/views/widgets/occasion_products_grid_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -154,50 +153,8 @@ class _OccasionViewBodyState extends State<OccasionViewBody> {
                                                       ).noProductsAvailable,
                                                     ),
                                                   )
-                                                  : GridView.builder(
-                                                    itemCount: products.length,
-                                                    gridDelegate:
-                                                        const SliverGridDelegateWithFixedCrossAxisCount(
-                                                          crossAxisCount: 2,
-                                                          mainAxisSpacing: 17,
-                                                          crossAxisSpacing: 17,
-                                                          childAspectRatio:
-                                                              163 / 229,
-                                                        ),
-                                                    itemBuilder: (
-                                                      context,
-                                                      index,
-                                                    ) {
-                                                      return InkWell(
-                                                        onTap: (){
-                                                          Navigator.of(context).pushNamed(
-                                                            RouteNames.productDetails,
-                                                            arguments: products[index],
-                                                          );
-                                                        },
-                                                        child: CustomProductItems(
-                                                          title:
-                                                              products[index]
-                                                                  .title ??
-                                                              AppLocalizations.of(
-                                                                context,
-                                                              ).product,
-                                                          imgCover:
-                                                              products[index]
-                                                                  .imgCover ??
-                                                              "https://flower.elevateegy.com/uploads/fefa790a-f0c1-42a0-8699-34e8fc065812-cover_image.png",
-                                                          priceAfterDiscount:
-                                                              products[index]
-                                                                  .priceAfterDiscount ??
-                                                              0,
-                                                          price:
-                                                              products[index]
-                                                                  .price ??
-                                                              0,
-                                                        ),
-                                                      );
-                                                    },
-                                                  ),
+                                                  : 
+                                                  OccasionProductsGridView(products: products)
                                         );
                                       }).toList(),
                                 ),
@@ -213,3 +170,4 @@ class _OccasionViewBodyState extends State<OccasionViewBody> {
     );
   }
 }
+
